refactor: moved all the code from android to androidTest inside library.
