+++ library/src/androidTest/kotlin/edu/usf/cutr/otp/tampa/PlanApiUnitTest.kt
@@ -1,16 +1,14 @@
-package edu.usf.cutr.otp.demo.tampa
+package edu.usf.cutr.otp.tampa
 
 import edu.usf.cutr.otp.plan.api.PlanApi
 import edu.usf.cutr.otp.plan.model.Planner
 import edu.usf.cutr.otp.plan.model.RequestParameters
-import kotlinx.coroutines.ExperimentalCoroutinesApi
 import org.junit.Assert.assertEquals
 import org.junit.Test
 import java.util.concurrent.CountDownLatch
 
 class PlanApiUnitTest {
 
-    @ExperimentalCoroutinesApi
     @Test
     fun testPlanApi() {
         val planApi =
